Updated documentation.

+++ src/moving_platform.hpp
@@ -39,10 +39,47 @@ class Platform {
     enum class Direction { Forward, Backward };
 
   private:
+    /**
+     * @brief Moves the platform forward.
+     *
+     * Move the platform forward for *distanceMilliMeters*.
+     *
+     * @param[in] distanceMilliMeters The amount of millimeters the platform should move.
+     */
     void movePlatformForward(int32_t distanceMilliMeters);
+
+    /**
+     * @brief Moves the platform backward.
+     *
+     * Move the platform backward for *distanceMilliMeters*.
+     *
+     * @param[in] distanceMilliMeters The amount of millimeters the platform should move.
+     */
     void movePlatformBackward(int32_t distanceMilliMeters);
+
+    /**
+     * @brief Wait until a certain amount of rotations have passed.
+     *
+     * This function looks at the hall sensor and only returns after a certain amount of rotations of the hall sensor have passed.
+     * This means that you have to be carefull to not give the wrong amount of rotations, otherwise it's stuck in the loops forever.
+     *
+     * @param[in] rotations The amount of rotations the hall sensor has to rotate.
+     */
     void waitUntilRotations(int32_t rotations);
-    void caclulateNewCount(const int32_t &newA, const int32_t &newB, const int32_t &previousA, int32_t &count);
+
+    /**
+     * @brief Calculate the new amount of rotations.
+     *
+     * This function gets the new and old positions of the hall sensor and then figures out which direction the hall sensor is
+     * moving and changed the count accordingly.
+     *
+     * @param[in] newA The new value of pin A of the hall sensor.
+     * @param[in] newB The new value of pin B of the hall sensor.
+     * @param[in] previousA The old value of pin A of the hall sensor.
+     * @param[in] count The current amount of rotations.
+     * @param[out] count The updated amount of rotations.
+     */
+    void caclulateNewCount(const int8_t &newA, const int8_t &newB, const int8_t &previousA, int32_t &count);
 
   public:
     Platform(UARTLib::UARTConnection &serialCon, hwlib::pin_in &hallSensorPinA, hwlib::pin_in &hallSensorPinB);
